Add opaque pointer argument to iterator callbacks
Add opaque pointer argument to iterator callbacks, providing a facility
to pass context from the caller of the iteration to the callback and
back.  Because API/ABI breakage was unavoidable, and to prevent future
API proliferation, the _filtered variants have been removed.

This API change is not backwards compatible; the library version has
been raised to 0x01060504UL.